requirements.txt
 
Flask
dash
dash-bootstrap-components
pytesseract
pdf2image
Pillow
langchain 
langchain-community 
langchain-core
python-docx
Werkzeug
pytesseract
requests